Create new Command macro
