+++ process/drivers.rs
@@ -6,27 +6,23 @@
 
 use std::{
     collections::{hash_map::Entry, HashMap},
-    env,
     fmt::Debug,
-    path::Path,
     process::{ExitStatus, Output},
     sync::{Mutex, RwLock},
 };
 
 use blue_build_recipe::Recipe;
-use blue_build_utils::constants::{
-    COSIGN_PRIVATE_KEY, COSIGN_PRIV_PATH, COSIGN_PUB_PATH, IMAGE_VERSION_LABEL,
-};
+use blue_build_utils::constants::IMAGE_VERSION_LABEL;
 use clap::Args;
-use log::{debug, info, trace, warn};
-use miette::{bail, miette, Result};
+use log::{debug, info, trace};
+use miette::{miette, Result};
 use once_cell::sync::Lazy;
 use typed_builder::TypedBuilder;
 use uuid::Uuid;
 
 use self::{
     buildah_driver::BuildahDriver,
-    cosign_driver::{CosignDriver, VerifyType},
+    cosign_driver::CosignDriver,
     docker_driver::DockerDriver,
     github_driver::GithubDriver,
     gitlab_driver::GitlabDriver,
@@ -81,25 +77,25 @@ pub struct DriverArgs {
     /// Select which driver to use to build
     /// your image.
     #[builder(default)]
-    #[arg(short = 'B', long)]
+    #[arg(short = 'B', long, group = "drivers")]
     build_driver: Option<BuildDriverType>,
 
     /// Select which driver to use to inspect
     /// images.
     #[builder(default)]
-    #[arg(short = 'I', long)]
+    #[arg(short = 'I', long, group = "drivers")]
     inspect_driver: Option<InspectDriverType>,
 
     /// Select which driver to use to sign
     /// images.
     #[builder(default)]
-    #[arg(short = 'S', long)]
+    #[arg(short = 'S', long, group = "drivers")]
     signing_driver: Option<SigningDriverType>,
 
     /// Select which driver to use to run
     /// containers.
     #[builder(default)]
-    #[arg(short = 'R', long)]
+    #[arg(short = 'R', long, group = "drivers")]
     run_driver: Option<RunDriverType>,
 }
 
@@ -293,7 +289,7 @@ impl SigningDriver for Driver {
         match Self::get_signing_driver() {
             SigningDriverType::Cosign => CosignDriver::generate_key_pair(),
             SigningDriverType::Podman => todo!(),
-            SigningDriverType::Docker => todo!(),
+            SigningDriverType::Docker => DockerDriver::generate_key_pair(),
         }
     }
 
@@ -301,19 +297,23 @@ impl SigningDriver for Driver {
         match Self::get_signing_driver() {
             SigningDriverType::Cosign => CosignDriver::check_signing_files(),
             SigningDriverType::Podman => todo!(),
-            SigningDriverType::Docker => todo!(),
+            SigningDriverType::Docker => DockerDriver::check_signing_files(),
         }
     }
 
-    fn sign_images<S, T>(image_name: S, tag: Option<T>) -> Result<()>
-    where
-        S: AsRef<str>,
-        T: AsRef<str>,
-    {
+    fn sign(image_digest: &str, key_arg: Option<String>) -> Result<()> {
         match Self::get_signing_driver() {
-            SigningDriverType::Cosign => CosignDriver::sign_images(image_name, tag),
+            SigningDriverType::Cosign => CosignDriver::sign(image_digest, key_arg),
             SigningDriverType::Podman => todo!(),
-            SigningDriverType::Docker => todo!(),
+            SigningDriverType::Docker => DockerDriver::sign(image_digest, key_arg),
+        }
+    }
+
+    fn verify(image_name_tag: &str, verify_type: VerifyType) -> Result<()> {
+        match Self::get_signing_driver() {
+            SigningDriverType::Cosign => CosignDriver::verify(image_name_tag, verify_type),
+            SigningDriverType::Podman => todo!(),
+            SigningDriverType::Docker => DockerDriver::verify(image_name_tag, verify_type),
         }
     }
 
@@ -321,7 +321,7 @@ impl SigningDriver for Driver {
         match Self::get_signing_driver() {
             SigningDriverType::Cosign => CosignDriver::signing_login(),
             SigningDriverType::Podman => todo!(),
-            SigningDriverType::Docker => todo!(),
+            SigningDriverType::Docker => DockerDriver::signing_login(),
         }
     }
 }
@@ -408,101 +408,4 @@ impl CiDriver for Driver {
             CiDriverType::Github => GithubDriver::generate_image_name(recipe),
         }
     }
-}
-
-fn get_private_key(check_fn: impl FnOnce(String) -> Result<()>) -> Result<()> {
-    match (
-        Path::new(COSIGN_PUB_PATH).exists(),
-        env::var(COSIGN_PRIVATE_KEY).ok(),
-        Path::new(COSIGN_PRIV_PATH),
-    ) {
-        (true, Some(cosign_priv_key), _) if !cosign_priv_key.is_empty() => {
-            check_fn("env://COSIGN_PRIVATE_KEY".to_string())
-        }
-        (true, _, cosign_priv_key_path) if cosign_priv_key_path.exists() => {
-            check_fn(cosign_priv_key_path.display().to_string())
-        }
-        (true, _, _) => {
-            bail!(
-                "{}{}{}{}{}{}{}",
-                "Unable to find private/public key pair.\n\n",
-                format_args!("Make sure you have a `{COSIGN_PUB_PATH}` "),
-                format_args!("in the root of your repo and have either {COSIGN_PRIVATE_KEY} "),
-                format_args!("set in your env variables or a `{COSIGN_PRIV_PATH}` "),
-                "file in the root of your repo.\n\n",
-                "See https://blue-build.org/how-to/cosign/ for more information.\n\n",
-                "If you don't want to sign your image, use the `--no-sign` flag."
-            )
-        }
-        _ => Ok(()),
-    }
-}
-
-#[allow(clippy::needless_pass_by_value)]
-fn sign_images<S, T, Sign, Verify>(
-    image_name: S,
-    tag: Option<T>,
-    sign_fn: Sign,
-    verify_fn: Verify,
-) -> Result<()>
-where
-    S: AsRef<str>,
-    T: AsRef<str>,
-    Sign: Fn(&str, Option<String>) -> Result<()>,
-    Verify: Fn(&str, VerifyType) -> Result<()>,
-{
-    let image_name = image_name.as_ref();
-    let tag = tag.as_ref().map(AsRef::as_ref);
-    trace!("sign_images({image_name}, {tag:?}, sign_fn, verify_fn)");
-
-    let inspect_opts = GetMetadataOpts::builder().image(image_name);
-
-    let inspect_opts = if let Some(tag) = tag {
-        inspect_opts.tag(tag).build()
-    } else {
-        inspect_opts.build()
-    };
-
-    let image_digest = Driver::get_metadata(&inspect_opts)?.digest;
-    let image_name_tag = tag.map_or_else(|| image_name.to_owned(), |t| format!("{image_name}:{t}"));
-    let image_digest = format!("{image_name}@{image_digest}");
-
-    match (
-        Driver::get_ci_driver(),
-        // Cosign public/private key pair
-        env::var(COSIGN_PRIVATE_KEY),
-        Path::new(COSIGN_PRIV_PATH),
-    ) {
-        // Cosign public/private key pair
-        (_, Ok(cosign_private_key), _)
-            if !cosign_private_key.is_empty() && Path::new(COSIGN_PUB_PATH).exists() =>
-        {
-            sign_fn(
-                &image_digest,
-                Some(format!("--key=env://{COSIGN_PRIVATE_KEY}")),
-            )?;
-            verify_fn(&image_name_tag, VerifyType::File(COSIGN_PUB_PATH.into()))?;
-        }
-        (_, _, cosign_priv_key_path) if cosign_priv_key_path.exists() => {
-            sign_fn(
-                &image_digest,
-                Some(format!("--key={}", cosign_priv_key_path.display())),
-            )?;
-            verify_fn(&image_name_tag, VerifyType::File(COSIGN_PUB_PATH.into()))?;
-        }
-        // Gitlab keyless
-        (CiDriverType::Github | CiDriverType::Gitlab, _, _) => {
-            sign_fn(&image_digest, None)?;
-            verify_fn(
-                &image_name_tag,
-                VerifyType::Keyless {
-                    issuer: Driver::oidc_provider()?,
-                    identity: Driver::keyless_cert_identity()?,
-                },
-            )?;
-        }
-        _ => warn!("Not running in CI with cosign variables, not signing"),
-    }
-
-    Ok(())
 }
