+++ process/drivers/traits.rs
@@ -1,9 +1,16 @@
-use std::process::{ExitStatus, Output};
+use std::{
+    env,
+    path::Path,
+    process::{ExitStatus, Output},
+};
 
 use blue_build_recipe::Recipe;
-use log::{debug, info, trace};
+use blue_build_utils::constants::{COSIGN_PRIVATE_KEY, COSIGN_PRIV_PATH, COSIGN_PUB_PATH};
+use log::{debug, info, trace, warn};
 use miette::{bail, miette, Result};
 use semver::{Version, VersionReq};
+
+use crate::drivers::{types::CiDriverType, Driver};
 
 use super::{
     image_metadata::ImageMetadata,
@@ -154,6 +161,12 @@ pub trait RunDriver: Sync + Send {
     fn run_output(opts: &RunOpts) -> std::io::Result<Output>;
 }
 
+#[derive(Debug, Clone)]
+pub enum VerifyType {
+    File(String),
+    Keyless { issuer: String, identity: String },
+}
+
 pub trait SigningDriver {
     /// Generate a new private/public key pair.
     ///
@@ -168,6 +181,41 @@ pub trait SigningDriver {
     /// Will error if the files cannot be verified.
     fn check_signing_files() -> Result<()>;
 
+    /// Signs the image digest.
+    ///
+    /// ## The `key_arg`
+    /// The `key_arg` is expected to be in the format of the
+    /// `--key` argument on the `cosign` CLI.
+    ///
+    /// ### Examples
+    ///   `"--key=cosign.key"`
+    ///
+    ///   `"--key=env://[ENV_VAR]"`
+    ///
+    ///   `"--key=azurekms://[VAULT_NAME][VAULT_URI]/[KEY]"`
+    ///
+    ///   `"--key=awskms://[ENDPOINT]/[ID/ALIAS/ARN]"`
+    ///
+    ///   `"--key=gcpkms://projects/[PROJECT]/locations/global/keyRings/[KEYRING]/cryptoKeys/[KEY]/versions/[VERSION]"`
+    ///
+    ///   `"--key=hashivault://[KEY]"`
+    ///
+    ///   `"--key=k8s://[NAMESPACE]/[KEY]"`
+    ///
+    /// # Errors
+    /// Will error if signing fails.
+    fn sign(image_digest: &str, key_arg: Option<String>) -> Result<()>;
+
+    /// Verifies the image.
+    ///
+    /// The image can be verified either with `VerifyType::File` containing
+    /// the public key contents, or with `VerifyType::Keyless` containing
+    /// information about the `issuer` and `identity`.
+    ///
+    /// # Errors
+    /// Will error if the image fails to be verified.
+    fn verify(image_name_tag: &str, verify_type: VerifyType) -> Result<()>;
+
     /// Sign an image given the image name and tag.
     ///
     /// # Errors
@@ -175,13 +223,98 @@ pub trait SigningDriver {
     fn sign_images<S, T>(image_name: S, tag: Option<T>) -> Result<()>
     where
         S: AsRef<str>,
-        T: AsRef<str>;
+        T: AsRef<str>,
+    {
+        let image_name = image_name.as_ref();
+        let tag = tag.as_ref().map(AsRef::as_ref);
+        trace!("sign_images({image_name}, {tag:?}, sign_fn, verify_fn)");
+
+        let inspect_opts = GetMetadataOpts::builder().image(image_name);
+
+        let inspect_opts = if let Some(tag) = tag {
+            inspect_opts.tag(tag).build()
+        } else {
+            inspect_opts.build()
+        };
+
+        let image_digest = Driver::get_metadata(&inspect_opts)?.digest;
+        let image_name_tag =
+            tag.map_or_else(|| image_name.to_owned(), |t| format!("{image_name}:{t}"));
+        let image_digest = format!("{image_name}@{image_digest}");
+
+        match (
+            Driver::get_ci_driver(),
+            // Cosign public/private key pair
+            env::var(COSIGN_PRIVATE_KEY),
+            Path::new(COSIGN_PRIV_PATH),
+        ) {
+            // Cosign public/private key pair
+            (_, Ok(cosign_private_key), _)
+                if !cosign_private_key.is_empty() && Path::new(COSIGN_PUB_PATH).exists() =>
+            {
+                Self::sign(
+                    &image_digest,
+                    Some(format!("--key=env://{COSIGN_PRIVATE_KEY}")),
+                )?;
+                Self::verify(&image_name_tag, VerifyType::File(COSIGN_PUB_PATH.into()))?;
+            }
+            (_, _, cosign_priv_key_path) if cosign_priv_key_path.exists() => {
+                Self::sign(
+                    &image_digest,
+                    Some(format!("--key={}", cosign_priv_key_path.display())),
+                )?;
+                Self::verify(&image_name_tag, VerifyType::File(COSIGN_PUB_PATH.into()))?;
+            }
+            // Gitlab keyless
+            (CiDriverType::Github | CiDriverType::Gitlab, _, _) => {
+                Self::sign(&image_digest, None)?;
+                Self::verify(
+                    &image_name_tag,
+                    VerifyType::Keyless {
+                        issuer: Driver::oidc_provider()?,
+                        identity: Driver::keyless_cert_identity()?,
+                    },
+                )?;
+            }
+            _ => warn!("Not running in CI with cosign variables, not signing"),
+        }
+
+        Ok(())
+    }
 
     /// Runs the login logic for the signing driver.
     ///
     /// # Errors
     /// Will error if login fails.
     fn signing_login() -> Result<()>;
+}
+
+pub(super) fn get_private_key(check_fn: impl FnOnce(String) -> Result<()>) -> Result<()> {
+    match (
+        Path::new(COSIGN_PUB_PATH).exists(),
+        env::var(COSIGN_PRIVATE_KEY).ok(),
+        Path::new(COSIGN_PRIV_PATH),
+    ) {
+        (true, Some(cosign_priv_key), _) if !cosign_priv_key.is_empty() => {
+            check_fn("env://COSIGN_PRIVATE_KEY".to_string())
+        }
+        (true, _, cosign_priv_key_path) if cosign_priv_key_path.exists() => {
+            check_fn(cosign_priv_key_path.display().to_string())
+        }
+        (true, _, _) => {
+            bail!(
+                "{}{}{}{}{}{}{}",
+                "Unable to find private/public key pair.\n\n",
+                format_args!("Make sure you have a `{COSIGN_PUB_PATH}` "),
+                format_args!("in the root of your repo and have either {COSIGN_PRIVATE_KEY} "),
+                format_args!("set in your env variables or a `{COSIGN_PRIV_PATH}` "),
+                "file in the root of your repo.\n\n",
+                "See https://blue-build.org/how-to/cosign/ for more information.\n\n",
+                "If you don't want to sign your image, use the `--no-sign` flag."
+            )
+        }
+        _ => Ok(()),
+    }
 }
 
 /// Allows agnostic retrieval of CI-based information.
